requirements.txt
python-telegram-bot==20.7
fastapi
uvicorn
gspread
oauth2client
